feat(healthcheck): Store more information in case of HC fail
Closes gravitee-io/issues#813
